Fix a crash when Graph is deleted.
Deleting a Graph that is still computing in another thread could lead to
a crash because Graph::~Graph(), which stops the thread, is called after
the destructor for the subclass, which may destroy resources needed by
the thread. Now the right way to delete a Graph is to call
Graph::dispose(), which stops the thread first, then deletes it.