Bug fix: Infinite loop when closing the application.
When the last window is closed, Qt wants to end the application. It
starts by trying to stop QtConcurrent threads, which is not going to
happen until the Graphs get destroyed, leading to a deadlock. Now Graph
just uses QThread to do its work, instead of QtConcurrent::run, avoiding
this problem.